test: Move toBeIdenticalTo() matcher to CustomMatchersForTaskBuilder.ts
